move filebrowser package out of components package
... as "components" package should be used for jetpack compose components that can be shared for multiple screens
